updated instagram.py
Fixed bug that prevented spider from collected lount count values. The problem was that the key was incorrect. Instead 'edge_liked_by', I changed it to 'edge_media_preview_like' and it worked! No missing values!!

Take it for a spin and let me know if it worked for you.

Have a great weekend and thank you for this awesome program!